\name{swishdbtools-package}
\alias{swishdbtools-package}
\alias{swishdbtools}
\docType{package}
\title{
getPassword and connect2postgres
}
\description{
getPassword and connect2postgres~
}
\details{
\tabular{ll}{
Package: \tab swishdbtools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-12-13\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <ivan.hanigan@gmail.com>

}
\references{

}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
